+++ common/src/main/java/io/github/fishstiz/cursors_extended/mixin/WindowMixin.java
@@ -40,6 +40,9 @@ public abstract class WindowMixin {
             }
             return true;
         }
+        if (CursorsExtended.CONFIG.isAggressiveCursor()) {
+            return true;
+        }
         return cursor.handle() != cursors_extended$currentCursorHandle;
     }